refactor: adjust imports and update tests
