--- pytest_splunk_addon/standard_lib/cim_tests/__init__.py
@@ -1,11 +1,0 @@
-"""
-Test generation mechanism to verify the CIM compatibility of an Add-on 
-"""
-
-from .data_model import DataModel
-from .data_model_handler import DataModelHandler
-from .data_set import DataSet
-from .field_test_helper import FieldTestHelper
-from .json_schema import JSONSchema
-from .test_generator import CIMTestGenerator
-from .test_templates import CIMTestTemplates
