+++ pytest_splunk_addon/standard_lib/sample_generation/sample_xdist_generator.py
@@ -5,7 +5,7 @@ import pickle
 import pytest
 from filelock import FileLock
 
-from . import SampleGenerator
+from .sample_generator import SampleGenerator
 
 
 class SampleXdistGenerator:
